+++ src/Rhetos.CommonConcepts/Dom.DefaultConcepts/DataStructure/Writing/LoadOldItemsTakeCodeGenerator.cs
@@ -17,16 +17,12 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using Rhetos.Dsl.DefaultConcepts;
-using System.Globalization;
-using System.ComponentModel.Composition;
 using Rhetos.Compiler;
-using Rhetos.Extensibility;
+using Rhetos.DatabaseGenerator.DefaultConcepts;
 using Rhetos.Dsl;
+using Rhetos.Dsl.DefaultConcepts;
+using Rhetos.Extensibility;
+using System.ComponentModel.Composition;
 
 namespace Rhetos.Dom.DefaultConcepts
 {
@@ -34,18 +30,54 @@ namespace Rhetos.Dom.DefaultConcepts
     [ExportMetadata(MefProvider.Implements, typeof(LoadOldItemsTakeInfo))]
     public class LoadOldItemsTakeCodeGenerator : IConceptCodeGenerator
     {
+        private readonly IDslModel _dslModel;
+        private readonly ConceptMetadata _conceptMetadata;
+
+        public LoadOldItemsTakeCodeGenerator(IDslModel dslModel, ConceptMetadata conceptMetadata)
+        {
+            _dslModel = dslModel;
+            _conceptMetadata = conceptMetadata;
+        }
+
         public void GenerateCode(IConceptInfo conceptInfo, ICodeBuilder codeBuilder)
         {
             var info = (LoadOldItemsTakeInfo)conceptInfo;
-            codeBuilder.InsertCode(GetSnippet(info), LoadOldItemsCodeGenerator.SelectPropertiesTag, info.LoadOldItems);
+            string propertyName = info.GetPropertyName();
+
+            codeBuilder.InsertCode(
+                $",\r\n                {propertyName} = item.{info.Path}",
+                LoadOldItemsCodeGenerator.SelectPropertiesTag, info.LoadOldItems);
+
+            codeBuilder.InsertCode(
+                $"\r\n            public {GetCsPropertyType(info)} {propertyName} {{ get; set; }}",
+                LoadOldItemsCodeGenerator.OldItemPropertiesTag, info.LoadOldItems);
         }
 
-        private string GetSnippet(LoadOldItemsTakeInfo info)
+        /// <summary>
+        /// Returns the C# type of the property selected by the path, as declared in the entity's queryable class.
+        /// </summary>
+        private string GetCsPropertyType(LoadOldItemsTakeInfo info)
         {
-            return string.Format(
-@",
-                {0} = item.{1}",
-                info.GetPropertyName(), info.Path);
+            var property = DslUtility.GetPropertyByPath(info.LoadOldItems.SaveMethod.Entity, info.Path, _dslModel);
+            if (property.IsError)
+                throw new DslConceptSyntaxException(info, "Invalid path: " + property.Error);
+
+            // A path that ends with a reference property name selects the navigation property.
+            // A path that ends with the reference's ID property resolves to GuidPropertyInfo instead.
+            if (property.Value is ReferencePropertyInfo reference)
+            {
+                if (!DslUtility.IsQueryable(reference.Referenced))
+                    throw new DslConceptSyntaxException(info, $"The path ends with {reference.GetUserDescription()}," +
+                        $" but there is no navigation property because {reference.Referenced.GetUserDescription()} is not queryable." +
+                        $" Use the reference ID property instead ('{info.Path}ID').");
+                return $"Common.Queryable.{reference.Referenced.Module.Name}_{reference.Referenced.Name}";
+            }
+
+            string csPropertyType = _conceptMetadata.GetCsPropertyType(property.Value);
+            if (string.IsNullOrEmpty(csPropertyType))
+                throw new DslConceptSyntaxException(info, $"{property.Value.GetKeywordOrTypeName()} is not supported" +
+                    $" for {info.GetKeywordOrTypeName()}, because it does not provide concept metadata for C# property type.");
+            return csPropertyType;
         }
     }
 }