+++ test/Rhetos.CommonConcepts.Test/Tools/PluginsContainerMock.cs
@@ -1,0 +1,43 @@
+/*
+    Copyright (C) 2014 Omega software d.o.o.
+
+    This file is part of Rhetos.
+
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU Affero General Public License as
+    published by the Free Software Foundation, either version 3 of the
+    License, or (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU Affero General Public License for more details.
+
+    You should have received a copy of the GNU Affero General Public License
+    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+*/
+
+using Rhetos.Extensibility;
+using System;
+using System.Collections.Generic;
+
+namespace Rhetos.CommonConcepts.Test.Mocks
+{
+    public class PluginsContainerMock<T> : IPluginsContainer<T>
+    {
+        private readonly IReadOnlyCollection<T> _plugins;
+
+        public PluginsContainerMock(params T[] plugins)
+        {
+            _plugins = plugins;
+        }
+
+        public IReadOnlyCollection<T> GetPlugins() => _plugins;
+
+        public Type GetMetadata(T plugin, string metadataKey) => null;
+
+        public Type GetMetadata(Type pluginType, string metadataKey) => null;
+
+        public IReadOnlyCollection<T> GetImplementations(Type activator) => Array.Empty<T>();
+    }
+}